Libraries cleanup
