+++ HID-Emulator/platformio.ini
@@ -1,5 +1,7 @@
 [common]
 build_flags = -O3 -Wall -D__DEBUG__
+              -DRESTFULL_UI
+              -DRESTFULL_CALLS
               '-DFIRMWARE="HID-Emulator 17.05.0"'
 lib_deps = ArduinoJson
            WebSockets
