+++ IOTbase/platformio.ini
@@ -1,6 +1,8 @@
 [common]
 build_flags = -O3 -Wall -D__DEBUG__
-              '-DFIRMWARE="IOT 17.07"'
+              -DRESTFULL_UI
+              -DRESTFULL_CALLS
+              '-DFIRMWARE="IOT 17.07.0"'
 lib_deps = NeoPixelBus
            ArduinoJson
            WebSockets
