+++ IOTbase/src/iot.cpp
@@ -3,10 +3,10 @@
 
 #include <commons.h>
 #include <file_system.h>
-#include <service_rest.h>
 #include <service_ota.h>
-#include <service_led_strip.h>
-#include <configuration.h>
+#include <service_rest_robust.h>
+#include <service_io_restfull.h>
+#include <service_led_strip_persist.h>
 
 std::vector<Service *> services;
 
@@ -21,9 +21,9 @@ void ICACHE_FLASH_ATTR setup() {
                 *host_name = ConfigJSON::getString(CONFIG_GLOBAL_JSON, {"host-name"});
         services.push_back(Log::getInstance());
         services.push_back(OtaService::get_instance(admin_pass));
-        LedStripService *led_strip = new PersistentLedStripService();
-        services.push_back(init_rest(new RestService(admin_acc, admin_pass, 80), led_strip, ALL));
-        services.push_back((Service *) led_strip);
+        services.push_back(new RestServiceRobust(admin_acc, admin_pass, 80, ALL));
+        RestFullIO::register_callbacks((RestService *) services.back());
+        services.push_back(new PersistentLedStripService((RestService *) services.back()));
         Log::println("Credentials: [%s:%s]", admin_acc, admin_pass);
         if (MDNS.begin(host_name)) {
             Log::println("Hostname: [%s]", host_name);
@@ -33,7 +33,7 @@ void ICACHE_FLASH_ATTR setup() {
         checked_free(host_name);
     }
     wifi_config_reset();
-    Devices::parse_devices(CONFIG_IO_JSON, DIGITAL_IO);
+    RestFullIO::parse_devices(CONFIG_IO_JSON, DIGITAL_IO);
     delay(500);
 }
