+++ lib/service_ntp_time/service_ntp_time.h
@@ -5,7 +5,7 @@
 #include <Ticker.h>
 #include <WiFiUdp.h>
 #include <commons.h>
-#include <logger.h>
+#include <service_log.h>
 
 typedef enum {
     UTC__12_00 = -720,
